requirements.txt
pandas 
matplotlib 
streamlit 
mysql-connector-python
